test : PathFinder 테스트를 PathServiceTest와 통합한다
